requirements.txt
pyzmq
click
tox